# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VELOXA API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /owners/{token}:
    get:
      operationId: getOwnerByToken
      tags: [owners]
      summary: Get owner profile by NFC token or username
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Owner profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
        "404":
          description: Not found
  /owners/{id}/profile:
    patch:
      operationId: updateOwnerProfile
      tags: [owners]
      summary: Update owner profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerUpdate"
      responses:
        "200":
          description: Updated owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
...
  /me:
    get:
      operationId: getMe
      tags: [me]
      summary: Get logged-in owner profile
      responses:
        "200":
          description: Owner profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
        "401":
          description: Not authenticated
  /me/profile:
    patch:
      operationId: updateMyProfile
      tags: [me]
      summary: Update logged-in owner profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    Owner:
      type: object
      required: [id, username, name]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        title:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        linkedin:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        socialLinks:
          type: ["object", "null"]
        createdAt:
          type: string
...
    OwnerUpdate:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        company:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        instagram:
          type: string
        linkedin:
          type: string
        city:
          type: string
        industry:
          type: string
        socialLinks:
          type: object
...