# lib/api-zod/src/generated/types/owner.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VELOXA API
 * OpenAPI spec version: 0.1.0
 */
import type { OwnerSocialLinks } from './ownerSocialLinks';

export interface Owner {
  id: number;
  username: string;
  name: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  socialLinks?: OwnerSocialLinks;
  createdAt?: string;
}